refactor index-buffers to std::vector<uint32_t> where applicable refactor line primitive creation ( performance, see  PrimitiveModeNormalsTest.gltf sample) add gltf points primitive to 3ds Max mesh creation path slightly improve readability of CreateMaxNode(...) implement related helper functionality

+++ HSglTFImporter/SourceFiles/DracoDecode.cpp
@@ -47,7 +47,7 @@
 //=======================================================================
 // Create vertex ID list from compressed buffer
 //=======================================================================
-void GetDracoMeshIndexList(cgltf_buffer_view* bufferView, std::vector<float> &tbl)
+void GetDracoMeshIndexList(cgltf_buffer_view* bufferView, std::vector<uint32_t> &tbl)
 {
 	tbl.clear();
